requirements.txt
streamlit
groq
python-dotenv
discord.py
pydub
google-cloud-texttospeech
google-generativeai
gTTS
audio-recorder-streamlit
